Updated Order model and database
I created a remote database and changed the connection string. I also made some changes in the Order model and updated the database schema to ensure the data integrity in the orders table. I added an order dto class and used instead of an order and added an order confirmation page.